FIXUP: drm/msm: Add SDM845 DPU support
Create fixup.

Kcr-depends-on-sha: 25fdd5933e4c0f5fe2ea5cd59994f8ac5fbe90ef

+++ drivers/gpu/drm/msm/disp/dpu1/dpu_hw_catalog.h
@@ -19,7 +19,7 @@
  */
 #define MAX_BLOCKS    12
 
-#define DPU_HW_VER(MAJOR, MINOR, STEP) (((MAJOR & 0xF) << 28)    |\
+#define DPU_HW_VER(MAJOR, MINOR, STEP) ((((unsigned int)MAJOR & 0xF) << 28)    |\
 		((MINOR & 0xFFF) << 16)  |\
 		(STEP & 0xFFFF))
